requirements.txt
Flask==2.3.3
qrcode[pil]==7.4.2
Pillow==10.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
